[F] edit name and handler in dataSource
