+++ packages/frontend/src/presenter/DataSources/constants.ts
@@ -23,33 +23,3 @@ export const INITIAL_STATE = {
   handler: {},
   transforms: [],
 };
-
-export const ODATA_BATCHES = [
-  { label: '-', value: '' },
-  { label: 'multipart', value: 'multipart' },
-  { label: 'json', value: 'json' },
-];
-
-export const THRIFT_PROTOCOLS = [
-  { label: 'Binary', value: 'binary' },
-  { label: 'Compact', value: 'compact' },
-  { label: 'Json', value: 'json' },
-];
-
-export enum SOURCE_NAMES {
-  FHIR = 'fhir',
-  GRAPHQL = 'graphql',
-  GRPC = 'grpc',
-  JSON_SCHEMA = 'jsonSchema',
-  MONGOOSE = 'mongoose',
-  MYSQL = 'mysql',
-  NEO4J = 'neo4j',
-  ODATA = 'odata',
-  OPEN_API = 'openapi',
-  POST_GRAPHILE = 'postgraphile',
-  SOAP = 'soap',
-  THRIFT = 'thrift',
-  TUQL = 'tuql',
-  CONTENTFUL_HANDLER = 'ContentfulHandler',
-  SLACK_HANDLER = 'SlackHandler',
-}
