+++ packages/frontend/src/view/DataSources/constants.ts
@@ -36,3 +36,15 @@ export const GRAPHQL_METHODS = [
   { label: 'GET', value: 'GET' },
   { label: 'POST', value: 'POST' },
 ];
+
+export const ODATA_BATCHES = [
+  { label: '-', value: '' },
+  { label: 'multipart', value: 'multipart' },
+  { label: 'json', value: 'json' },
+];
+
+export const THRIFT_PROTOCOLS = [
+  { label: 'Binary', value: 'binary' },
+  { label: 'Compact', value: 'compact' },
+  { label: 'Json', value: 'json' },
+];
